requirements.txt
numpy==1.21.2
opencv-python==4.5.3.56
Pillow==8.2
tqdm==4.53.0
matplotlib==3.2.1
scipy==1.8.0
open3d==0.13.0
Shapely==1.7.1
torch==1.9.0
torchvision==0.10.0 
torchaudio==0.9.0
